// service/src/utils/types.ts
import { Conversation, Friend, FriendRequest, Group, GroupMessage, Message, User } from "./typeorm";
import { Request } from 'express';
export type CreateUserDetails = {
    username: string;
    password: string;
    firstName: string;
    lastName: string;
};

export type ValidateUserDetails = {
    username: string;
    password: string;
};

export type FindUserParams = Partial<{
    id: number;
    email: string;
    username: string;
}>;

export type CreateConversationParams = {
    username: string;
    message: string;
};

export type ConversationIdentityType = 'author' | 'recipient';

export type FindParticipantParams = Partial<{
    id: number;
}>;

export interface AuthenticatedRequest extends Request {
    user: User;
}

export type CreateParticipantParams = {
    id: number;
};

export type CreateMessageParams = {
    content: string;
    id: number;
    user: User;
};

export type CreateMessageResponse = {
    message: Message;
    conversation: Conversation;
};

export type DeleteMessageParams = {
    userId: number;
    conversationId: number;
    messageId: number;
}

export type EditMessageParams = {
    conversationId: number;
    messageId: number;
    userId: number;
    content: string;
}

export type CreateGroupParam = {
    creator: User;
    title?: string;
    users: string[];
}

export type FetchGroupsParams = {
    userId: number;
}

export type FindUserOptions = Partial<{
    selectAll: boolean;
}>;

export type CreateGroupMessageParams = {
    groupId: number;
    content: string;
    author: User;
}


export type CreateGroupMessageResponse = {
    message: GroupMessage;
    group: Group;
}

export type DeleteGroupMessageParams = {
    userId: number;
    groupId: number;
    messageId: number;
};

export type EditGroupMessageParams = {
    groupId: number;
    messageId: number;
    userId: number;
    content: string;
}
export type AddGroupRecipientParams = {
    id: number;
    username: string;
    userId: number;
}

export type RemoveGroupRecipientParams = {
    id: number;
    removeUserId: number;
    issuerId: number;
}


export type AddGroupUserResponse = {
    group: Group;
    user: User;
}


export type RemoveGroupUserResponse = {
    group: Group;
    user: User;

}


export type AccessParams = {
    id: number;
    userId: number;
}

export type TransferOwerParams = {
    userId: number;
    groupId: number;
    newOwnerId: number;
}

export type LeaveGroupPrams = {
    id: number;
    userId: number;
}

export type CheckUserGroupParams = {
    id: number;
    userId: number;
}

export type CreateFriendParams = {
    user: User;
    username: string;

}

export type FriendRequestStatus = 'accepted' | 'pending' | 'rejected';

export type FriendRequestParams = {
    id: number;
    userId: number;
}

export type CancelFriendRequestParams = {
    id: number;
    userId: number;
}

export type DeleteFriendRequestParams = {
    id: number;
    userId: number;
}

export type AcceptFriendRequestResponse = {
    friend: Friend;
    friendRequest: FriendRequest
}

export type RemovedFriendEventPayload = {
    friend: Friend;
    userId: number;
}

export type UserProfiles = Partial<{
    banner: Express.Multer.File[];
    avatar: Express.Multer.File[];
}>;

export type UpdateUserProfileParams = Partial<{
    about: string;
    banner: Express.Multer.File;
    avatar: Express.Multer.File;
}>;

export type UploadImageParams = {
    key: string;
    file: Express.Multer.File;
}

export type GetConversationMessagesParams = {
    id: number;
    limit: number;
}

export type UpdateConversationParams = {
    id: number;
    lastMessageSent: Message;
}


export type UpdatePresenceStatus = 'online' | 'away' | 'offline' | 'dnd';
export type UpdateStatusMessageParams = {
    user: User;
    statusMessage: string;
}